generate_prompt_and_art/requirements.txt
grpcio==1.48.1
openai==0.11.0
Pillow==9.4.0
protobuf==4.21.12
pytest==7.2.1
python-dotenv==0.21.1
requests==2.28.1
setuptools==65.6.3
